requirements.txt
aniso8601==3.0.2
awscli==1.15.71
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.1
boto3==1.7.70
botocore==1.10.70
certifi==2018.4.16
chardet==3.0.4
cheroot==6.4.0
CherryPy==17.0.0
click==6.7
docutils==0.14
Flask==1.0.2
Flask-Admin==1.5.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
jaraco.functools==1.20
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
more-itertools==4.3.0
numpy==1.15.0
portend==2.3
PyGeoj==0.22
pyshp==1.2.12
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
raven==6.9.0
requests==2.19.1
s3transfer==0.1.13
Shapely==1.6.4.post2
simplekml==1.3.0
six==1.11.0
tempora==1.13
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
visvalingamwyatt==0.1.2
